requirements.txt
berserk==0.10.0
pyTelegramBotAPI==3.7.7
python-lichess==0.9